Move create_empty_array() to statevars.py
